update brand image

+++ src/main/java/uz/carapp/rentcarapp/repository/AttachmentRepository.java
@@ -3,5 +3,5 @@ package uz.carapp.rentcarapp.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.carapp.rentcarapp.domain.Attachment;
 
-public interface AttachmentRepository extends JpaRepository<Attachment, Integer> {
+public interface AttachmentRepository extends JpaRepository<Attachment, Long> {
 }
